refactor: extract snapshot reconciliation helpers from chat-view
Move pure utility functions for snapshot reconciliation, timeline
manipulation, and action matching into chat-snapshot-helpers.ts.
